% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bindRows}
\alias{bindRows}
\title{Bind Rows of List Elements}
\usage{
bindRows(input_list)
}
\arguments{
\item{input_list}{}
}
\value{
Dataframe of each list element as a row
}
\description{
Bind Rows of List Elements
}
