% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_lists_with_name}
\alias{find_lists_with_name}
\title{Search a Nested List to Get a List and its Respective Elements}
\usage{
find_lists_with_name(nested_list, target_name)
}
\arguments{
\item{nested_list}{a list with nested elements of lists}

\item{target_name}{identifier for list(s)}
}
\value{
list with elements all lists matching the target_name
}
\description{
Search a Nested List to Get a List and its Respective Elements
}
\examples{
nested_list_example <- list(
teamA = list(
offense = list(
  quarterback = "Patrick Mahomes",
  running_back = "Isiah Pacheco",
  wide_receivers = c("Rashee Rice", "Hollywood Brown")
),
defense = list(
  linebacker = "Nick Bolton",
  cornerbacks = c("Trent McDuffie", "Joshua Williams")
)
),
teamB = list(
  offense = list(
    quarterback = "Josh Allen",
    running_back = "James Cook",
    wide_receivers = c("Stefon Diggs", "Khalil Shakir")
  ),
  defense = list(
    linebacker = "Matt Milano",
    cornerbacks = c("Kaiir Elam", "Taron Johnson")
  )
)
)
find_lists_with_name(nested_list_example, "offense")
}
