% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_variable_in_nested_list}
\alias{find_variable_in_nested_list}
\title{Return a Value in a Nested List by Searching for Its Name}
\usage{
find_variable_in_nested_list(nested_list, target_variable_name)
}
\arguments{
\item{nested_list}{List with elements that are nested lists}

\item{target_variable_name}{The name of the element to get the value}
}
\value{
Value associated with name
}
\description{
Return a Value in a Nested List by Searching for Its Name
}
