% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_config.R
\name{setup_league_meta}
\alias{setup_league_meta}
\title{Internal: Retrieve sport metadata for a league key}
\usage{
setup_league_meta(league_key, settings_type)
}
\description{
Internal: Retrieve sport metadata for a league key
}
\keyword{internal}
