% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{yf_game_from_key}
\alias{yf_game_from_key}
\title{Get Specific Game Data for Select Seasons using Game Key}
\usage{
yf_game_from_key(game_key)
}
\arguments{
\item{game_key}{Identifier for a fantasy sport's season}
}
\value{
Data frame of metadata for input game keys
}
\description{
Get Specific Game Data for Select Seasons using Game Key
}
\examples{
nfl_461 <- yf_game_from_key("461")
}
