% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{yf_game_from_year}
\alias{yf_game_from_year}
\title{Get Specific Game Data for Select Seasons using Year}
\usage{
yf_game_from_year(year)
}
\arguments{
\item{year}{Season year for fantasy sport}
}
\value{
Data frame of metadata for input years
}
\description{
Get Specific Game Data for Select Seasons using Year
}
\examples{
nfl_2025 <- yf_game_from_year(2025)
}
