% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{yf_gamekey_current_season}
\alias{yf_gamekey_current_season}
\title{Get the Game Key for the Current Fantasy Season}
\usage{
yf_gamekey_current_season()
}
\value{
Game Key id for current season
}
\description{
Get the Game Key for the Current Fantasy Season
}
\examples{
current_seasons_gamekey <- yf_gamekey_current_season()
}
