% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{yf_gamekeys}
\alias{yf_gamekeys}
\title{Get Game Keys for All Available Fantasy Seasons}
\usage{
yf_gamekeys(game_code)
}
\value{
Data frame of all game metadata
}
\description{
Get Game Keys for All Available Fantasy Seasons
}
\examples{
all_nfl_seasons <- yf_gamekeys()
}
