% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{yf_get}
\alias{yf_get}
\title{GET Yahoo Fantasy API endpoints}
\usage{
yf_get(endpoint)
}
\arguments{
\item{endpoint}{API endpoint url to access resources and subresources}
}
\value{
List from endpoint response
}
\description{
GET Yahoo Fantasy API endpoints
}
\examples{
# fantasy_football_seasons <- yf_get("games;game_codes=nfl")
}
