% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{yf_league}
\alias{yf_league}
\title{Get League Metadata}
\usage{
yf_league(league_key)
}
\arguments{
\item{league_key}{Identifier for league}
}
\value{
Dataframe of league metadata
}
\description{
Get League Metadata
}
\examples{
league <- yf_league("461.l.442310")
}
