% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{yf_league_scoreboard}
\alias{yf_league_scoreboard}
\title{Get the Scoreboard for All Teams}
\usage{
yf_league_scoreboard(league_key, week = NULL)
}
\arguments{
\item{league_key}{Identifier for league.}

\item{week}{Week number of league}
}
\value{
Dataframe
}
\description{
Get the Scoreboard for All Teams
}
\examples{
yf_league_scoreboard("466.l.278123", week = 1)
}
