% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{yf_user_leagues}
\alias{yf_user_leagues}
\title{Get All Leagues You Participate In}
\usage{
yf_user_leagues()
}
\value{
Data frame with a league in each row
}
\description{
Get All Leagues You Participate In
}
\examples{
my_leagues <- yf_user_leagues()
}
