% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{yf_league_settings}
\alias{yf_league_settings}
\title{Get League Settings}
\usage{
yf_league_settings(league_key)
}
\arguments{
\item{league_key}{Identifier for league}
}
\value{
Dataframe of league settings
}
\description{
Get League Settings
}
\examples{
league_roster_settings <- yf_league_settings("461.l.442310")
}
